Dump in two 1024-byte blocks to recover all ROM
The MC6805P-series (P2. P4. P6) have varying amounts of internal ROM,
and a boot ROM at the end of ROM.

This commit reads all 2048 bytes of the address space, 1024 bytes at a
time. The CPU is reset and resynchronised, then 1024 bytes of data are
read, and printed as a C array.

There is an issue in that the 6805Px is a dynamic design and has a
minimum clock frequency: dumping then outputting over serial stops the
clock and corrupts the internal registers. This is why we have to do a
resynchronisation at the start of each 1K block.

The final block contains all 2048 bytes of memory space, including SFRs,
RAM, User ROM (for any device in the 6805P series), and the Motorola
bootloader/selfcheck code.